requirements.txt
kurigram==2.2.18
telethon
motor==3.6.0
pytgcrypto
python-dotenv
watchdog
typer
httpx


